Fix TitleBlockZen circular types
